+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 2.0.2 January 04, 2024
 
-- Changed Standards to ask user whether to fix warnings when running phpcbf ([DESCW-1744](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-1744))
+- Changed Standards::phpcbf to fix all warnings and errors by default ([DESCW-1744](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-1744))
 
 ### 2.0.1 October 25, 2023